# api_definition.yaml
openapi: 3.0.0
info:
  title: spot-share
  version: '1.0'
servers:
  - url: 'http://spot-share.nolit.net'
  - url: 'http://localhost'
    description: ''
paths:
...
  /routes:
    get:
      summary: Get Routes
      tags: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SimpleRoute'
      operationId: get-routes
    post:
      summary: Post a Route with routeDetails
      operationId: post-routes
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  routeId:
                    type: integer
                  routeName:
                    type:
                      - string
                      - 'null'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                routeName:
                  type:
                    - string
                    - 'null'
                routeDetails:
                  type: array
                  items:
                    $ref: '#/components/schemas/RouteDetail'
        description: ''
  '/routes/{id}':
    parameters:
      - schema:
          type: string
        name: id
        in: path
        required: true
    get:
      summary: Get a Route with routeDetails
      tags: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  routeId:
                    type: integer
                  routeName:
                    type:
                      - string
                      - 'null'
                  routeDetails:
                    type: array
                    items:
                      type: object
                      properties:
                        routeDetailId:
                          type: integer
                        spot:
                          $ref: '#/components/schemas/Spot'
                        scheduledDateTime:
                          type:
                            - string
                            - 'null'
                        orderNumber:
                          type:
                            - integer
                            - 'null'
                        favoritePoint:
                          type: integer
                        memo:
                          type: string
      operationId: get-routes-id
    delete:
      summary: Delete a route with routeDetails
      operationId: delete-routes-id
      responses:
        '200':
          description: OK
    patch:
      summary: Patch a Route with routeDetails
      operationId: patch-routes-id
      responses:
        '200':
          description: OK
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                routeName:
                  type: string
                routeDetails:
                  type: array
                  items:
                    $ref: '#/components/schemas/RouteDetail'
...
components:
  schemas:
    Spot:
      title: Spot
      type: object
      properties:
        spotId:
          type: integer
        country:
          $ref: '#/components/schemas/Country'
        spotName:
          type: string
        url:
          type:
            - string
            - 'null'
        summary:
          type:
            - string
            - 'null'
        imagePaths:
          type: array
          items:
            type: string
        costExpectation:
          type:
            - integer
            - 'null'
        requiredHours:
          type:
            - integer
            - 'null'
        requiredMinutes:
          type:
            - integer
            - 'null'
      required:
        - spotId
    SimpleRoute:
      title: RouteResponse
      type: object
      properties:
        routeId:
          type: integer
        routeName:
          type: string
    Country:
      title: Country
      type: object
      properties:
        code:
          type: string
        countryName:
          type: string
        currency:
          type:
            - string
            - 'null'
        timezone:
          type:
            - string
            - 'null'
    RouteDetail:
      title: RouteDetail
      type: object
      properties:
        spotId:
          type: integer
        orderNumber:
          type:
            - integer
            - 'null'
        scheduledDateTime:
          type:
            - string
            - 'null'
        favoritePoint:
          type:
            - string
            - 'null'
        memo:
          type: string
        routeDetailId:
          type:
            - integer
            - 'null'